refactor: extract session setup module
